[Popover] Quick refactorisation (#5798)
