+++ src/Popover/Popover.js
@@ -6,7 +6,7 @@ import propTypes from '../utils/propTypes';
 import Paper from '../Paper';
 import throttle from 'lodash.throttle';
 import PopoverAnimationDefault from './PopoverAnimationDefault';
-import {isIOS, getOffsetTop} from '../utils/isIOS';
+import {isIOS, getOffsetTop} from '../utils/iOSHelpers';
 
 class Popover extends Component {
   static propTypes = {
@@ -243,6 +243,9 @@ class Popover extends Component {
     };
 
     a.right = rect.right || a.left + a.width;
+
+    // The fixed positioning isn't respected on iOS when an input is focused.
+    // We need to compute the position from the top of the page and not the viewport.
     if (isIOS() && document.activeElement.tagName === 'INPUT') {
       a.bottom = getOffsetTop(el) + a.height;
     } else {
@@ -361,24 +364,28 @@ class Popover extends Component {
 
     if (targetPosition.top < 0 || targetPosition.top + target.bottom > window.innerHeight) {
       let newTop = anchor[anchorPos.vertical] - target[positions.y[0]];
-      if (newTop + target.bottom <= window.innerHeight)
+      if (newTop + target.bottom <= window.innerHeight) {
         targetPosition.top = Math.max(0, newTop);
-      else {
+      } else {
         newTop = anchor[anchorPos.vertical] - target[positions.y[1]];
-        if (newTop + target.bottom <= window.innerHeight)
+        if (newTop + target.bottom <= window.innerHeight) {
           targetPosition.top = Math.max(0, newTop);
+        }
       }
     }
+
     if (targetPosition.left < 0 || targetPosition.left + target.right > window.innerWidth) {
       let newLeft = anchor[anchorPos.horizontal] - target[positions.x[0]];
-      if (newLeft + target.right <= window.innerWidth)
+      if (newLeft + target.right <= window.innerWidth) {
         targetPosition.left = Math.max(0, newLeft);
-      else {
+      } else {
         newLeft = anchor[anchorPos.horizontal] - target[positions.x[1]];
-        if (newLeft + target.right <= window.innerWidth)
+        if (newLeft + target.right <= window.innerWidth) {
           targetPosition.left = Math.max(0, newLeft);
+        }
       }
     }
+
     return targetPosition;
   }
