+++ src/utils/iOSHelpers.js
